# docs/openapi.yaml
openapi: 3.1.0
info:
  title: InkFlow CRM API
  description: |
    REST API for InkFlow tattoo studio platform.

    Base path: /api (server context-path).

    Response envelope: ApiResponse<T> - see docs/API_FORMAT.md.

    Auth:
    - CRM (admin mobile / web): Supabase JWT in Authorization Bearer header.
      Custom claims: tenant_id, role, location_ids (via Supabase hook).
    - Consumer (B2C app): Supabase JWT for /public/consumer/** routes.
    - Public reads: GET /public/catalog/** and GET /public/artists/** - no token.
  contact:
    name: InkFlow
    email: dev@inkflow.studio
  license:
    name: Proprietary
  version: 1.0.0
servers:
- url: /api
  description: Current host (context-path included)
paths:
...
  /payments/refund:
    post:
      tags:
      - CRM Â· Payments
      operationId: processRefund
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessRefundRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePaymentDto"
  /payments/process:
    post:
      tags:
      - CRM Â· Payments
      operationId: processPayment
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessPaymentRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePaymentDto"
  /payments/monobank/webhook:
    post:
      tags:
      - System Â· Webhooks
      operationId: webhook
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MonobankWebhookPayload"
        required: true
      responses:
        "200":
          description: OK
  /payments/monobank/invoice:
    post:
      tags:
      - System Â· Webhooks
      operationId: createInvoice
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOnlineInvoiceRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseOnlineInvoiceDto"
...
  /leaves:
    get:
      tags:
      - CRM Â· Leave
      operationId: getAllLeaves
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
      - name: leaveType
        in: query
        required: false
        schema:
          type: string
      - name: from
        in: query
        required: false
        schema:
          type: string
      - name: to
        in: query
        required: false
        schema:
          type: string
      - name: locationId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: staffIds
        in: query
        required: false
        schema:
          type: array
          items:
            type: string
            format: uuid
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListLeaveRequestDto"
    post:
      tags:
      - CRM Â· Leave
      operationId: createLeave
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeaveRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseLeaveRequestDto"
...
  /staff/{id}:
    get:
      tags:
      - CRM Â· Staff
      operationId: getStaff
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseStaffDetailDto"
    delete:
      tags:
      - CRM Â· Staff
      operationId: deleteStaff
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
    patch:
      tags:
      - CRM Â· Staff
      operationId: updateStaff
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStaffRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseStaffDto"
...
  /leaves/{id}/status:
    patch:
      tags:
      - CRM Â· Leave
      operationId: updateLeaveStatus
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeaveStatusRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseLeaveRequestDto"
  /leaves/{id}/cancel:
    patch:
      tags:
      - CRM Â· Leave
      operationId: cancelLeave
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseLeaveRequestDto"
...
  /staff/{id}/google/status:
    get:
      tags:
      - CRM Â· Integrations
      operationId: getStatus
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseGoogleCalendarStatusResponse"
...
  /payments/appointment/{appointmentId}:
    get:
      tags:
      - CRM Â· Payments
      operationId: getAppointmentPayments
      parameters:
      - name: appointmentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListPaymentDto"
  /payments/appointment/{appointmentId}/summary:
    get:
      tags:
      - CRM Â· Payments
      operationId: getPaymentSummary
      parameters:
      - name: appointmentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseAppointmentPaymentSummaryDto"
  /leaves/{id}:
    get:
      tags:
      - CRM Â· Leave
      operationId: getLeaveById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseLeaveRequestDto"
    delete:
      tags:
      - CRM Â· Leave
      operationId: deleteLeave
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
  /leaves/staff/{staffId}:
    get:
      tags:
      - CRM Â· Leave
      operationId: getLeavesByStaffId
      parameters:
      - name: staffId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: startDate
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: endDate
        in: query
        required: false
        schema:
          type: string
          format: date
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListLeaveRequestDto"
  /leaves/staff/{staffId}/check:
    get:
      tags:
      - CRM Â· Leave
      operationId: checkStaffOnLeave
      parameters:
      - name: staffId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: date
        in: query
        required: true
        schema:
          type: string
          format: date
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseBoolean"
  /leaves/pending-count:
    get:
      tags:
      - CRM Â· Leave
      operationId: getPendingCount
      parameters:
      - name: locationId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseLong"
  /leaves/calendar:
    get:
      tags:
      - CRM Â· Leave
      operationId: getLeavesForCalendar
      parameters:
      - name: from
        in: query
        required: true
        schema:
          type: string
          format: date
      - name: to
        in: query
        required: true
        schema:
          type: string
          format: date
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListLeaveRequestDto"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/ErrorDto"
...
    PaginationDto:
      type: object
      properties:
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
...
    ApiResponseVoid:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        pagination:
          $ref: "#/components/schemas/PaginationDto"
        error:
          $ref: "#/components/schemas/ErrorResponse"
...
    ApiResponseStaffDto:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/StaffDto"
        pagination:
          $ref: "#/components/schemas/PaginationDto"
        error:
          $ref: "#/components/schemas/ErrorResponse"
...
    ProcessRefundRequest:
      type: object
      properties:
        transactionId:
          type: string
          format: uuid
        amount:
          type: number
          minimum: 0.01
        reason:
          type: string
        paymentMethod:
          type: string
      required:
      - amount
      - paymentMethod
      - reason
      - transactionId
    ApiResponsePaymentDto:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/PaymentDto"
        pagination:
          $ref: "#/components/schemas/PaginationDto"
        error:
          $ref: "#/components/schemas/ErrorResponse"
    PaymentDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        paymentType:
          type: string
        paymentTypeLabel:
          type: string
        amount:
          type: number
        paymentMethod:
          type: string
        paymentMethodLabel:
          type: string
        cashAmount:
          type: number
        cardAmount:
          type: number
        tipAmount:
          type: number
        description:
          type: string
        date:
          type: string
          format: date-time
        receiptNumber:
          type: string
        isRefunded:
          type: boolean
        refundedAmount:
          type: number
        refundableAmount:
          type: number
        originalTransactionId:
          type: string
          format: uuid
        refundReason:
          type: string
        processedById:
          type: string
          format: uuid
        processedByName:
          type: string
        clientName:
          type: string
        createdAt:
          type: string
          format: date-time
    ProcessPaymentRequest:
      type: object
      properties:
        appointmentId:
          type: string
          format: uuid
        amount:
          type: number
          minimum: 0.01
        paymentMethod:
          type: string
        paymentType:
          type: string
        cashAmount:
          type: number
        cardAmount:
          type: number
        tipAmount:
          type: number
        description:
          type: string
      required:
      - amount
      - appointmentId
      - paymentMethod
    MonobankWebhookPayload:
      type: object
      properties:
        amountDecimal:
          type: number
        invoiceId:
          type: string
        status:
          type: string
        amount:
          type: integer
          format: int64
        ccy:
          type: integer
          format: int32
        finalAmount:
          type: integer
          format: int64
        createdDate:
          type: string
        modifiedDate:
          type: string
        reference:
          type: string
        paymentInfo:
          $ref: "#/components/schemas/PaymentInfo"
    PaymentInfo:
      type: object
      properties:
        feeDecimal:
          type: number
        maskedPan:
          type: string
        approvalCode:
          type: string
        rrn:
          type: string
        paymentSystem:
          type: string
        country:
          type: string
        fee:
          type: integer
          format: int64
    CreateOnlineInvoiceRequest:
      type: object
      properties:
        appointmentId:
          type: string
          format: uuid
        amount:
          type: number
          minimum: 1.0
        paymentType:
          type: string
          pattern: ^(service_payment|deposit)$
      required:
      - amount
      - appointmentId
    ApiResponseOnlineInvoiceDto:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/OnlineInvoiceDto"
        pagination:
          $ref: "#/components/schemas/PaginationDto"
        error:
          $ref: "#/components/schemas/ErrorResponse"
    OnlineInvoiceDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        appointmentId:
          type: string
          format: uuid
        monobankInvoiceId:
          type: string
        amount:
          type: number
        pageUrl:
          type: string
        status:
          type: string
        paymentType:
          type: string
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    CreateLeaveRequest:
      type: object
      properties:
        staffId:
          type: string
          format: uuid
        leaveType:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        reason:
          type: string
        notes:
          type: string
      required:
      - endDate
      - leaveType
      - staffId
      - startDate
    ApiResponseLeaveRequestDto:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/LeaveRequestDto"
        pagination:
          $ref: "#/components/schemas/PaginationDto"
        error:
          $ref: "#/components/schemas/ErrorResponse"
    LeaveRequestDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        staffId:
          type: string
          format: uuid
        staffName:
          type: string
        staffAvatar:
          type: string
        leaveType:
          type: string
        status:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        daysCount:
          type: integer
          format: int32
        reason:
          type: string
        notes:
          type: string
        approvedById:
          type: string
          format: uuid
        approvedByName:
          type: string
        approvedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    UpdateStaffRequest:
      type: object
      properties:
        firstName:
          type: string
          maxLength: 50
          minLength: 1
        lastName:
          type: string
          maxLength: 50
          minLength: 1
        email:
          type: string
        phone:
          type: string
        avatar:
          type: string
        calendarColor:
          type: string
          pattern: "^#[0-9A-Fa-f]{6}$"
        specialization:
          type: array
          items:
            type: string
        portfolioImages:
          type: array
          items:
            type: string
        bio:
          type: string
        locationIds:
          type: array
          items:
            type: string
            format: uuid
        status:
          type: string
          pattern: ^(working|on_vacation|sick_leave|fired)$
        salaryType:
          type: string
          pattern: ^(none|fixed|percent)$
        salaryRate:
          type: number
          minimum: 0.0
        bankDetails:
          type: string
        position:
          type: string
        birthday:
          type: string
          format: date
        taxId:
          type: string
        iban:
          type: string
        bankCard:
          type: string
        availableForOnlineBooking:
          type: boolean
        isServiceProvider:
          type: boolean
        instagram:
          type: string
        hourlyRate:
          type: number
        dontDoList:
          type: array
          items:
            type: string
...
    UpdateLeaveStatusRequest:
      type: object
      properties:
        status:
          type: string
        notes:
          type: string
      required:
      - status
...
    ApiResponseStaffDetailDto:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/StaffDetailDto"
        pagination:
          $ref: "#/components/schemas/PaginationDto"
        error:
          $ref: "#/components/schemas/ErrorResponse"
...
    StaffDetailDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        avatar:
          type: string
        role:
          type: string
        calendarColor:
          type: string
        specialization:
          type: array
          items:
            type: string
        portfolioImages:
          type: array
          items:
            type: string
        bio:
          type: string
        status:
          type: string
        accountStatus:
          type: string
        locations:
          type: array
          items:
            $ref: "#/components/schemas/LocationDto"
        schedule:
          type: array
          items:
            $ref: "#/components/schemas/ScheduleDto"
        stats:
          $ref: "#/components/schemas/StaffStatsDto"
        salaryType:
          type: string
        salaryRate:
          type: number
        bankDetails:
          type: string
        isServiceProvider:
          type: boolean
        instagram:
          type: string
        hourlyRate:
          type: number
        dontDoList:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ApiResponseGoogleCalendarStatusResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/GoogleCalendarStatusResponse"
        pagination:
          $ref: "#/components/schemas/PaginationDto"
        error:
          $ref: "#/components/schemas/ErrorResponse"
    GoogleCalendarStatusResponse:
      type: object
      properties:
        connected:
          type: boolean
        email:
          type: string
        configured:
          type: boolean
...
    ApiResponseListPaymentDto:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/PaymentDto"
        pagination:
          $ref: "#/components/schemas/PaginationDto"
        error:
          $ref: "#/components/schemas/ErrorResponse"
    ApiResponseAppointmentPaymentSummaryDto:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/AppointmentPaymentSummaryDto"
        pagination:
          $ref: "#/components/schemas/PaginationDto"
        error:
          $ref: "#/components/schemas/ErrorResponse"
    AppointmentPaymentSummaryDto:
      type: object
      properties:
        appointmentId:
          type: string
          format: uuid
        servicePrice:
          type: number
        discount:
          type: number
        finalPrice:
          type: number
        totalPaid:
          type: number
        depositPaid:
          type: number
        servicePaid:
          type: number
        totalRefunded:
          type: number
        totalTips:
          type: number
        remainingBalance:
          type: number
        isFullyPaid:
          type: boolean
        hasDeposit:
          type: boolean
        hasRefunds:
          type: boolean
        payments:
          type: array
          items:
            $ref: "#/components/schemas/PaymentDto"
        paymentStatus:
          type: string
...
    ApiResponseListLeaveRequestDto:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/LeaveRequestDto"
        pagination:
          $ref: "#/components/schemas/PaginationDto"
        error:
          $ref: "#/components/schemas/ErrorResponse"
    ApiResponseBoolean:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: boolean
        pagination:
          $ref: "#/components/schemas/PaginationDto"
        error:
          $ref: "#/components/schemas/ErrorResponse"
    ApiResponseLong:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: integer
          format: int64
        pagination:
          $ref: "#/components/schemas/PaginationDto"
        error:
          $ref: "#/components/schemas/ErrorResponse"
...
  securitySchemes:
    CrmBearer:
      type: http
      description: Supabase JWT for CRM staff (OWNER/ADMIN/ARTIST).
      scheme: bearer
      bearerFormat: JWT
    ConsumerBearer:
      type: http
      description: Supabase JWT for B2C consumer app users.
      scheme: bearer
      bearerFormat: JWT