# src/main/java/com/inkflow/crm/module/monobank/controller/MonobankController.java
package com.inkflow.crm.module.monobank.controller;

import com.inkflow.crm.domain.enums.Permission;
import com.inkflow.crm.common.dto.ApiResponse;
import com.inkflow.crm.module.monobank.dto.*;
import com.inkflow.crm.module.monobank.service.MonobankService;
import com.inkflow.crm.security.RequirePermission;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;
import io.swagger.v3.oas.annotations.tags.Tag;

@Slf4j
@RestController
@RequestMapping("/payments/monobank")
@RequiredArgsConstructor
@Tag(name = "System · Webhooks")
public class MonobankController {

    private final MonobankService monobankService;

    @PostMapping("/invoice")
    @RequirePermission(Permission.PAYMENTS_PROCESS)
    public ResponseEntity<ApiResponse<OnlineInvoiceDto>> createInvoice(
            @Valid @RequestBody CreateOnlineInvoiceRequest request) {
        OnlineInvoiceDto invoice = monobankService.createInvoice(request);
        log.info("Monobank invoice created via API: appointmentId={}", request.getAppointmentId());

        return ResponseEntity.ok(ApiResponse.success(invoice));
    }

    @PostMapping("/webhook")
    public ResponseEntity<Void> webhook(@RequestBody MonobankWebhookPayload payload) {
        log.info("Monobank webhook: invoiceId={} status={}", payload.getInvoiceId(), payload.getStatus());
        monobankService.handleWebhook(payload);
        return ResponseEntity.ok().build();
    }
}